Move create_sparsity_pattern to own file
